chore: Update React dependencies
This commit updates the React dependencies in the package.json file. It adds the "react-bootstrap" package with version "^2.10.2" as a dependency. This update is necessary to incorporate the use of React Bootstrap components in the project.

Note: The commit message has been generated based on the provided code changes and recent user commits.