feat: Update AddCtrl component to use relative paths for images
This commit updates the AddCtrl component to use relative paths for the images. The src attribute of the images has been modified to "../assets/add.png" and "../assets/addgroup.png" respectively. This change ensures that the images are correctly displayed when the component is rendered.

Note: The commit message has been generated based on the provided code changes and recent user commits.