feat: Add History component to App
This commit adds the History component to the App component in order to display the transaction history. The import statement for History has been added and the component has been included in the JSX code. This change enhances the functionality of the application by providing a way to view past transactions.

Note: The commit message has been generated based on the provided code changes and recent user commits.