feat: Add findByGroupId method to GroupRepository
This commit adds a new method, findByGroupId, to the GroupRepository interface in the GroupRepository.java file. This method allows for retrieving a GroupCase object based on its groupId. This change enhances the functionality of the repository by providing an additional query method.

Note: The commit message has been generated based on the provided code changes and recent user commits.