feat: Add Account to newTransfer and newGroup objects
This commit adds the `Account` property to the `newTransfer` and `newGroup` objects in the `handleSignupClick` functions of the `AddTransfer` and `CreateGroup` components respectively. This ensures that the `Account` value is included when creating a new transfer or group.

Note: The commit message has been generated based on the provided code changes and recent user commits.