feat: Add AddID component to AddCtrl
This commit adds the AddID component to the AddCtrl file. The AddID component is imported and used in the AddCtrl component, allowing for the display and functionality of adding an ID. This change enhances the functionality of the AddCtrl component by providing the ability to add an ID.

Note: The commit message has been generated based on the provided code changes and recent user commits.