refactor: Update waitForTransfer and waitForAccept endpoints
This commit updates the `waitForTransfer` and `waitForAccept` endpoints in the `ApiController` class. It retrieves the user information from the `myDataRepository` based on the provided account and performs some operations on the retrieved data. The response now includes the `groupName` and `money` fields, which contain the names and amounts of trades that are in the "waitPay" and "wait" states respectively. If there are no trades in these states, the fields are set to "null". This change improves the functionality and response of the endpoints.

Note: The commit message has been generated based on the provided code changes and recent user commits.